vertex_agents/cost_metrics_agent/requirements.txt
# Merged at deploy with google-cloud-aiplatform[adk,agent_engines] by the ADK CLI.
psycopg[binary]>=3.2
google-cloud-bigquery>=3.0.0
